Mathjax rendering was broken for some reason. Refactored, works now
I'm still not sure why or when it stopped working. But the problem
was clearly on our end. This refactor should make it much easier
to maintain the Mathjax code.

Also tested Mathjax v3, which also works, but unfortunately it
lacks support for siunitx.

Upgraded from v2.7.1 to v2.7.7 (both versions work with this
refactored code). 2.7.7 was the latest 2.x version that appears
to work with our enabled TeX extensions.